--- app/controllers/personal-details-get-names.ts
@@ -1,45 +1,0 @@
-import { NextFunction, Request, Response, Router } from 'express';
-import { paths } from '../paths';
-import { enterDetailsValidation } from '../utils/personal-details-schema';
-
-function getNamePage(req: Request, res: Response, next: NextFunction) {
-  try {
-    res.render('get-names-page.njk');
-  } catch (e) {
-    next(e);
-  }
-}
-
-function postNamePage(req: Request, res: Response, next: NextFunction) {
-  try {
-    const validation = enterDetailsValidation(req.body);
-    let errors = null;
-    if (validation) {
-      errors = validation;
-      res.render('get-names-page.njk', {
-        errors: {
-          errorList: errors,
-          fieldErrors: { givenNames: { text: errors[0].text }, familyName: { text: errors[1].text } }
-        }
-      });
-    } else {
-      res.render('get-names-page.njk', { familyName: req.body.familyName, givenNames: req.body.givenNames });
-    }
-  } catch (e) {
-    next(e);
-  }
-
-}
-
-function setupPersonalDetailsController(deps?: any): Router {
-  const router = Router();
-  router.get(paths.enterName, getNamePage);
-  router.post(paths.enterName, postNamePage);
-  return router;
-}
-
-export {
-  setupPersonalDetailsController,
-  getNamePage,
-  postNamePage
-};
