+++ app/routes.ts
@@ -4,8 +4,7 @@ import { setupHealthController } from './controllers/health';
 import { setupHomeOfficeDetailsController } from './controllers/home-office-details';
 import { setupIdamController } from './controllers/idam';
 import { setupIndexController } from './controllers/index';
-import { setupPersonalDetailsController } from './controllers/personal-details-get-names';
-import { setupDateOfBirthController } from './controllers/personal-details/date-of-birth';
+import { setupPersonalDetailsController } from './controllers/personal-details';
 import { setupTaskListController } from './controllers/task-list';
 import { setupTypeOfAppealController } from './controllers/type-of-appeal';
 
@@ -16,11 +15,9 @@ const healthController = setupHealthController();
 const idamController = setupIdamController();
 const taskListController = setupTaskListController();
 const homeOfficeDetailsController = setupHomeOfficeDetailsController();
-const personalDetailsController = setupPersonalDetailsController();
 const typeOfAppealController = setupTypeOfAppealController();
 const devNextPageController = setupDevNextPageController();
-const dateOfBirthController = setupDateOfBirthController();
-
+const personalDetailsController = setupPersonalDetailsController();
 // not protected by idam
 router.use(healthController);
 
@@ -32,6 +29,6 @@ router.use(homeOfficeDetailsController);
 router.use(personalDetailsController);
 router.use(typeOfAppealController);
 router.use(devNextPageController);
-router.use(dateOfBirthController);
+router.use(personalDetailsController);
 
 export { router };
