+++ views/appeal-application/appellant-names-page.njk
@@ -1,4 +1,4 @@
-{% extends 'layout.njk'  %}
+{% extends '../layout.njk'  %}
 
 {%  block content %}
     {{ errorSummary(errors) }}
